+++ admin/frontend/next.config.js
@@ -1,7 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   output: "standalone",
-  reactStrictMode: true
+  reactStrictMode: true,
+  async rewrites() {
+    return [
+      {
+        source: "/api/:path*",
+        destination: `${process.env.BACKEND_URL ?? "http://backend:3001"}/api/:path*`
+      }
+    ];
+  }
 };
 
 module.exports = nextConfig;